requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Parameter Efficient Fine-Tuning
peft>=0.6.0
# bitsandbytes>=0.41.0  # Uncomment for GPU quantization (Linux/WSL only)

# Evaluation Metrics
evaluate>=0.4.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Hyperparameter Optimization
optuna>=3.3.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.14.0

# Bias Detection
textblob>=0.17.1
# spacy>=3.6.0  # Optional - not required for core functionality

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipywidgets>=8.1.0

# Inference Interface
gradio>=4.0.0

# Web Scraping (optional - for custom data collection)
beautifulsoup4>=4.12.0
requests>=2.31.0
# newspaper3k>=0.2.8  # Optional